+++ NHS Prase/Model/Persistent Model/Machine/_PRHospital.h
@@ -35,23 +35,19 @@ extern const struct PRHospitalFetchedProperties {
 
 
 
-
-@property (nonatomic, strong) NSNumber* id;
-
+@property (nonatomic, strong) NSNumber *id;
 
 
-@property int64_t idValue;
-- (int64_t)idValue;
-- (void)setIdValue:(int64_t)value_;
+@property long long idValue;
+- (long long)idValue;
+- (void)setIdValue:(long long)value_;
 
 //- (BOOL)validateId:(id*)value_ error:(NSError**)error_;
 
 
 
 
-
-@property (nonatomic, strong) NSString* name;
-
+@property (nonatomic, strong) NSString *name;
 
 
 //- (BOOL)validateName:(id*)value_ error:(NSError**)error_;
@@ -60,17 +56,16 @@ extern const struct PRHospitalFetchedProperties {
 
 
 
-@property (nonatomic, strong) PRTrust *trust;
+@property (nonatomic, strong) PRTrust* trust;
 
 //- (BOOL)validateTrust:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSSet *wards;
+@property (nonatomic, strong) NSSet* wards;
 
 - (NSMutableSet*)wardsSet;
-
 
 
 
@@ -92,8 +87,8 @@ extern const struct PRHospitalFetchedProperties {
 - (NSNumber*)primitiveId;
 - (void)setPrimitiveId:(NSNumber*)value;
 
-- (int64_t)primitiveIdValue;
-- (void)setPrimitiveIdValue:(int64_t)value_;
+- (long long)primitiveIdValue;
+- (void)setPrimitiveIdValue:(long long)value_;
