+++ NHS Prase/Model/Persistent Model/Machine/_PRPMOSQuestion.h
@@ -37,28 +37,24 @@ extern const struct PRPMOSQuestionFetchedProperties {
 
 
 
-
-@property (nonatomic, strong) NSNumber* localizationID;
-
+@property (nonatomic, strong) NSNumber *localizationID;
 
 
-@property int64_t localizationIDValue;
-- (int64_t)localizationIDValue;
-- (void)setLocalizationIDValue:(int64_t)value_;
+@property long long localizationIDValue;
+- (long long)localizationIDValue;
+- (void)setLocalizationIDValue:(long long)value_;
 
 //- (BOOL)validateLocalizationID:(id*)value_ error:(NSError**)error_;
 
 
 
 
-
-@property (nonatomic, strong) NSNumber* questionID;
-
+@property (nonatomic, strong) NSNumber *questionID;
 
 
-@property int64_t questionIDValue;
-- (int64_t)questionIDValue;
-- (void)setQuestionIDValue:(int64_t)value_;
+@property long long questionIDValue;
+- (long long)questionIDValue;
+- (void)setQuestionIDValue:(long long)value_;
 
 //- (BOOL)validateQuestionID:(id*)value_ error:(NSError**)error_;
 
@@ -66,24 +62,23 @@ extern const struct PRPMOSQuestionFetchedProperties {
 
 
 
-@property (nonatomic, strong) NSOrderedSet *answerSets;
+@property (nonatomic, strong) NSOrderedSet* answerSets;
 
 - (NSMutableOrderedSet*)answerSetsSet;
 
 
 
 
-@property (nonatomic, strong) PRPMOS *pmos;
+@property (nonatomic, strong) PRPMOS* pmos;
 
 //- (BOOL)validatePmos:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) NSSet *questions;
+@property (nonatomic, strong) NSSet* questions;
 
 - (NSMutableSet*)questionsSet;
-
 
 
 
@@ -110,8 +105,8 @@ extern const struct PRPMOSQuestionFetchedProperties {
 - (NSNumber*)primitiveLocalizationID;
 - (void)setPrimitiveLocalizationID:(NSNumber*)value;
 
-- (int64_t)primitiveLocalizationIDValue;
-- (void)setPrimitiveLocalizationIDValue:(int64_t)value_;
+- (long long)primitiveLocalizationIDValue;
+- (void)setPrimitiveLocalizationIDValue:(long long)value_;
 
 
 
@@ -119,8 +114,8 @@ extern const struct PRPMOSQuestionFetchedProperties {
 - (NSNumber*)primitiveQuestionID;
 - (void)setPrimitiveQuestionID:(NSNumber*)value;
 
-- (int64_t)primitiveQuestionIDValue;
-- (void)setPrimitiveQuestionIDValue:(int64_t)value_;
+- (long long)primitiveQuestionIDValue;
+- (void)setPrimitiveQuestionIDValue:(long long)value_;
