+++ NHS Prase/Model/Persistent Model/Machine/_PRQuestion.h
@@ -43,14 +43,12 @@ extern const struct PRQuestionFetchedProperties {
 
 
 
-
-@property (nonatomic, strong) NSNumber* answerID;
-
+@property (nonatomic, strong) NSNumber *answerID;
 
 
-@property int64_t answerIDValue;
-- (int64_t)answerIDValue;
-- (void)setAnswerIDValue:(int64_t)value_;
+@property long long answerIDValue;
+- (long long)answerIDValue;
+- (void)setAnswerIDValue:(long long)value_;
 
 //- (BOOL)validateAnswerID:(id*)value_ error:(NSError**)error_;
 
@@ -58,52 +56,51 @@ extern const struct PRQuestionFetchedProperties {
 
 
 
-@property (nonatomic, strong) PRConcern *concern;
+@property (nonatomic, strong) PRConcern* concern;
 
 //- (BOOL)validateConcern:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) PRConcern *concernAsPrevent;
+@property (nonatomic, strong) PRConcern* concernAsPrevent;
 
 //- (BOOL)validateConcernAsPrevent:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) PRConcern *concernAsSerious;
+@property (nonatomic, strong) PRConcern* concernAsSerious;
 
 //- (BOOL)validateConcernAsSerious:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) PRNote *goodNote;
+@property (nonatomic, strong) PRNote* goodNote;
 
 //- (BOOL)validateGoodNote:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) PRNote *note;
+@property (nonatomic, strong) PRNote* note;
 
 //- (BOOL)validateNote:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) PRPMOSQuestion *pmosQuestion;
+@property (nonatomic, strong) PRPMOSQuestion* pmosQuestion;
 
 //- (BOOL)validatePmosQuestion:(id*)value_ error:(NSError**)error_;
 
 
 
 
-@property (nonatomic, strong) PRRecord *record;
+@property (nonatomic, strong) PRRecord* record;
 
 //- (BOOL)validateRecord:(id*)value_ error:(NSError**)error_;
-
 
 
 
@@ -120,8 +117,8 @@ extern const struct PRQuestionFetchedProperties {
 - (NSNumber*)primitiveAnswerID;
 - (void)setPrimitiveAnswerID:(NSNumber*)value;
 
-- (int64_t)primitiveAnswerIDValue;
-- (void)setPrimitiveAnswerIDValue:(int64_t)value_;
+- (long long)primitiveAnswerIDValue;
+- (void)setPrimitiveAnswerIDValue:(long long)value_;
